'version11Alert'

+++ src/main/java/com/glsid/medapp/controller/MedecinController.java
@@ -43,6 +43,9 @@ public class MedecinController {
 	
 	@GetMapping(path = "/medecins/deleteMedecins")
 	public String delete(Long id, String page, String size) {
+		if(!medecinRepository.findById(id).get().getListConsultations().isEmpty()) {
+			return "medecin/alert";
+		}
 		medecinRepository.deleteById(id);
 		return "redirect:/medecin/listeMedecin?page="+page+"&size="+size;
 	}
